+++ src/main/java/com/baichen/rpc/registry/ServiceRegistry.java
@@ -1,0 +1,12 @@
+package com.baichen.rpc.registry;
+
+import java.util.List;
+
+public interface ServiceRegistry {
+
+    void init(ServiceRegistryConfig config) throws Exception;
+
+    void registry(ServiceMateData service) throws Exception;
+
+    List<ServiceMateData> fetchSeviceList(String serviceName) throws Exception;
+}